Inventory in battle + item drops
Added support for inventory items to be used in battle by accessing the inventory array in Battle, along with adding a function to the inventory script that allows item removal to allow items to be consumed. Also added two more items, Metal and Goo, to the item list and added support for the items to be dropped at random after defeating an enemy, with dynamic text to print the items rewarded.